Update genetracks to be compatible with drawsvg 2.
Includes extra changes related to #938.
Also split SvgDiffer to a separate file, to make live coding easier.
